Crate akka_projection_rs_grpc

Akka projections with gRPC communication

Sources events over a gRPC connection, using the Akka GRPC protocol.

Modules

consumer

producer

proto

Structs

BadEvent

Declares a gRPC event cannot be mapped to an event envelope.

Envelope

An envelope for all types of gRPC events.

EventEnvelope

An envelope type that wraps a gRPC event associated with a specific entity.

NoMessage

Declares that a protobuf criteria is unable to be converted due to there being no message.

NotAnEventEnvelope

Returned when the envelope cannot be represented as an event envelope.

Type Aliases

OriginId

Identifies an event producer to a consumer

StreamId

The logical stream identifier, mapped to a specific internal entity type by the producer settings