Module akka_projection_rs_grpc::proto

Modules

consume_event_in

Nested message and enum types in ConsumeEventIn.

consume_event_out

Nested message and enum types in ConsumeEventOut.

event_consumer_service_client

Generated client implementations.

event_consumer_service_server

Generated server implementations.

event_producer_service_client

Generated client implementations.

event_producer_service_server

Generated server implementations.

filter_criteria

Nested message and enum types in FilterCriteria.

load_event_response

Nested message and enum types in LoadEventResponse.

stream_in

Nested message and enum types in StreamIn.

stream_out

Nested message and enum types in StreamOut.

Structs

ConsumeEventIn

ConsumeEventOut

ConsumerEventAck

ConsumerEventInit

must be the first event from the connecting producer, the producer must then not start emitting events until it sees a ConsumerEventStart

ConsumerEventStart

EntityIdOffset

Event

EventTimestampRequest

Retrieve the timestamp of a specific event.

EventTimestampResponse

Response to EventTimestampRequest.

ExcludeEntityIds

Exclude events for entities with the given entity ids, unless there is a matching include filter that overrides the exclude.

ExcludeRegexEntityIds

Exclude events for entities with entity ids matching the given regular expressions, unless there is a matching include filter that overrides the exclude.

ExcludeTags

Exclude events with any of the given tags, unless there is a matching include filter that overrides the exclude.

FilterCriteria

FilterReq

Add filter criteria to exclude and include events for matching entities.

FilteredEvent

Events that are filtered out are represented by this placeholder to be able to track sequence numbers without holes.

IncludeEntityIds

Include events for entities with the given entity ids. A matching include overrides a matching exclude.

IncludeRegexEntityIds

Include events for entities with entity ids matching the given regular expressions. A matching include overrides a matching exclude.

IncludeTags

Include events with any of the given tags. A matching include overrides a matching exclude.

IncludeTopics

Include events with any of the given matching topics. A matching include overrides a matching exclude.

InitReq

The first message must always be this InitReq to setup the stream. It can only be used as the first message.

KeepAlive

LoadEventRequest

Lazy loading of a specific event. Used when payload for a backtracking event is needed.

LoadEventResponse

Response to LoadEventRequest.

Offset

PersistenceIdSeqNr

RemoveExcludeEntityIds

Remove a previously added ExcludeEntityIds.

RemoveExcludeRegexEntityIds

Remove a previously added ExcludeRegexEntityIds.

RemoveExcludeTags

Remove a previously added ExcludeTags.

RemoveIncludeEntityIds

Remove a previously added IncludeEntityIds.

RemoveIncludeRegexEntityIds

Remove a previously added IncludeRegexEntityIds.

RemoveIncludeTags

Remove a previously added IncludeTags.

RemoveIncludeTopics

Remove a previously added IncludeTopics.

ReplayReq

Replay events for given entities.

ReplicaInfo

Used for Replicated Event Sourcing to filter events based on origin. For edge topologies, like star topologies, an edge replica is not connected to all other replicas, but should be able to receive events indirectly via the replica that it is consuming from.

StreamIn

StreamOut