Trait akka_projection_rs_grpc::proto::event_producer_service_server::EventProducerService

pub trait EventProducerService: Send + Sync + 'static {
    type EventsBySlicesStream: Stream<Item = Result<StreamOut, Status>> + Send + 'static;

    // Required methods
    fn events_by_slices<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Streaming<StreamIn>>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::EventsBySlicesStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn event_timestamp<'life0, 'async_trait>(
        &'life0 self,
        request: Request<EventTimestampRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<EventTimestampResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn load_event<'life0, 'async_trait>(
        &'life0 self,
        request: Request<LoadEventRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<LoadEventResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with EventProducerServiceServer.

Required Associated Types

type EventsBySlicesStream: Stream<Item = Result<StreamOut, Status>> + Send + 'static

Server streaming response type for the EventsBySlices method.

Required Methods

fn events_by_slices<'life0, 'async_trait>( &'life0 self, request: Request<Streaming<StreamIn>> ) -> Pin<Box<dyn Future<Output = Result<Response<Self::EventsBySlicesStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn event_timestamp<'life0, 'async_trait>( &'life0 self, request: Request<EventTimestampRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<EventTimestampResponse>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Used in some edge cases by the offset store to retrieve the timestamp for a certain event.

fn load_event<'life0, 'async_trait>( &'life0 self, request: Request<LoadEventRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<LoadEventResponse>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Lazy loading of a specific event.